+++ src/commands/full.rs
@@ -12,6 +12,7 @@ pub fn full(args: &RunArgsWithTest) -> Result<ExitCode> {
     let ignores = config.ignores()?;
     let repo = Repo::load();
     let stacks = stacks::discover_all(&ignores);
-    let tests = to_sequences(config.select_tests(&args.test)?);
+    let requested = config.tests_for(&args.test, |commands| commands.full.as_ref());
+    let tests = to_sequences(config.select_requested_tests(requested)?);
     run_tasks(&args.run, &config, &stacks, repo.as_ref(), tests)
 }